Modiyfing Resource Group Templates
The the naming convention for the resource group and tag names and values have been changed because of changes in requirements.

The parameter `OwnerDepartmentContent` has been changed to `OwnerSignInName` to reflect its intended purpose of denoting to whom the `Owner` role over the resource group should
be assigned.

Accompanying documentation has been changed.