Cleaning up comments in code.
This obtains a storage context based on the AZ credentials of Azure Runas Account AAD access of Storage Blob Containers is in preview mode. Allowing AAD access of Storage Blob Containers can only be set in the portal and `az-cli`. Powershell and ARM templates does not support this setting yet.

```powershell
$AZURE_STORAGE_CONTEXT = New-AzStorageContext -StorageAccountName "$AZURE_STORAGE_ACCOUNT" `
                                       -UseConnectedAccount
```