Cleaning up output stream.
The `Add-AzAccount` commandlet was returning an Azure AD context object to the output stream. Code has
been added to send that object to null.

This results in the JSON payload as the only object returned by runbook.

A minor change is that the hash table of the results of the ARM template deployment is expliclitly
cast as `PSCustomObject` prior to being piped to `ConvertTo-Json`.